+++ packages/embeds/embed-core/.gitignore
@@ -1,0 +1,1 @@
+src/tailwind.generated.css
