+++ packages/embeds/embed-core/package.json
@@ -7,7 +7,9 @@
     "build": "vite build",
     "build:cal": "NEXT_PUBLIC_WEBSITE_URL='https://cal.com' yarn build",
     "vite": "vite",
-    "dev": "run-p 'build --watch' 'vite --port 3002 --strict-port --open'",
+    "tailwind": "yarn tailwindcss -i ./src/styles.css -o ./src/tailwind.generated.css --watch",
+    "buildWatchAndServer":"run-p 'build --watch' 'vite --port 3002 --strict-port --open'",
+    "dev": "run-p 'tailwind' 'buildWatchAndServer'",
     "type-check": "tsc --pretty --noEmit",
     "lint": "eslint --ext .ts,.js src",
     "test-playwright": "yarn playwright test --config=playwright/config/playwright.config.ts",
@@ -26,6 +28,7 @@
   },
   "dependencies": {
     "autoprefixer": "^10.4.4",
+    "postcss": "^8.4.12",
     "tailwindcss": "^3.0.24"
   }
 }
