+++ packages/embeds/embed-core/tailwind.config.js
@@ -1,7 +1,10 @@
 module.exports = {
-  content: ["./src/**/*.ts"],
+  // purge: [],
+  content: ["**/*Html.ts"],
+  darkMode: false, // or 'media' or 'class'
   theme: {
     extend: {},
   },
+  variants: {},
   plugins: [],
 };
