+++ packages/embeds/embed-react/README.md
@@ -3,3 +3,9 @@
 Embed Cal Link as a React Component
 
 To know how to use it, follow the steps at <https://docs.cal.com/integrations/embed>
+
+TODO
+
+- Playwright tests.
+  - Need to what these tests should be as embed-core already have tests. We probably just need to verify that embed-core API is called appropriately.
+  - It would probably be better if Playwright tests exist at one place for all embeds.
